Refactor:DialogFragment defines an interface through which it delivers the events back to the host activity
https://developer.android.com/guide/topics/ui/dialogs.html#PassingEvents
